requirements.txt
starlette==0.13.4
torch==1.5.0
torchvision==0.6.0
Pillow==7.1.2
aiofiles==0.5.0
uvicorn==0.11.5
gunicorn==20.0.4